Extract model display mode enum.
